refactor(common): extract keys to own module
